first clean up process
