+++ numpyro/infer/__init__.py
@@ -2,15 +2,17 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from numpyro.infer.elbo import ELBO, RenyiELBO
-from numpyro.infer.mcmc import HMC, MCMC, NUTS, SA
-from numpyro.infer.svi import SVI
-from numpyro.infer.util import (
-    Predictive,
+from numpyro.infer.initialization import (
     init_to_feasible,
     init_to_median,
     init_to_prior,
     init_to_uniform,
     init_to_value,
+)
+from numpyro.infer.mcmc import HMC, MCMC, NUTS, SA
+from numpyro.infer.svi import SVI
+from numpyro.infer.util import (
+    Predictive,
     log_likelihood
 )
