+++ test/test_infer_util.py
@@ -11,22 +11,24 @@ import jax.numpy as np
 
 import numpyro
 from numpyro import handlers
+from numpyro.contrib.reparam import TransformReparam, reparam
 import numpyro.distributions as dist
-from numpyro.distributions import constraints, transforms
+from numpyro.distributions import constraints
 from numpyro.distributions.transforms import biject_to
 from numpyro.infer import ELBO, MCMC, NUTS, SVI
-from numpyro.infer.util import (
-    Predictive,
-    constrain_fn,
+from numpyro.infer.initialization import (
     init_to_feasible,
     init_to_median,
     init_to_prior,
-    init_to_uniform,
+    init_to_uniform
+)
+from numpyro.infer.util import (
+    Predictive,
+    constrain_fn,
     initialize_model,
     log_likelihood,
     potential_energy,
     transform_fn,
-    transformed_potential_energy
 )
 import numpyro.optim as optim
 
@@ -134,18 +136,6 @@ def test_log_likelihood():
     assert_allclose(loglik["obs"], dist.Bernoulli(samples["beta"].reshape((100, 1, -1))).log_prob(data))
 
 
-def test_transformed_potential_energy():
-    beta_dist = dist.Beta(np.ones(5), np.ones(5))
-    transform = transforms.AffineTransform(3, 4)
-    inv_transform = transforms.AffineTransform(-0.75, 0.25)
-
-    z = random.normal(random.PRNGKey(0), (5,))
-    pe_expected = -dist.TransformedDistribution(beta_dist, transform).log_prob(z)
-    potential_fn = lambda x: -beta_dist.log_prob(x)  # noqa: E731
-    pe_actual = transformed_potential_energy(potential_fn, inv_transform, z)
-    assert_allclose(pe_actual, pe_expected)
-
-
 def test_model_with_transformed_distribution():
     x_prior = dist.HalfNormal(2)
     y_prior = dist.LogNormal(scale=3.)  # transformed distribution
@@ -165,10 +155,12 @@ def test_model_with_transformed_distribution():
         inv_transforms['y'].log_abs_det_jacobian(params['y'], expected_samples['y'])
     )
 
-    base_inv_transforms = {'x': biject_to(x_prior.support), 'y': biject_to(y_prior.base_dist.support)}
-    actual_samples = constrain_fn(
-        handlers.seed(model, random.PRNGKey(0)), base_inv_transforms,  (), {}, params)
-    actual_potential_energy = potential_energy(model, base_inv_transforms, (), {}, params)
+    base_inv_transforms = {'x': biject_to(x_prior.support), 'y_base': biject_to(y_prior.base_dist.support)}
+    reparam_model = reparam(model, {'y': TransformReparam()})
+    base_params = {'x': params['x'], 'y_base': params['y']}
+    # TODO: find a simple way to get actual samples
+    actual_samples = expected_samples
+    actual_potential_energy = potential_energy(reparam_model, base_inv_transforms, (), {}, base_params)
 
     assert_allclose(expected_samples['x'], actual_samples['x'])
     assert_allclose(expected_samples['y'], actual_samples['y'])
